+++ internal/spotengine/exit_triggers_test.go
@@ -1,6 +1,7 @@
 package spotengine
 
 import (
+	"fmt"
 	"sync"
 	"testing"
 	"time"
@@ -69,6 +70,9 @@ func (s priceStubExchange) GetFundingFees(string, time.Time) ([]exchange.Funding
 }
 func (s priceStubExchange) GetClosePnL(string, time.Time) ([]exchange.ClosePnL, error) {
 	return nil, nil
+}
+func (s priceStubExchange) GetWithdrawFee(string, string) (float64, error) {
+	return 0, fmt.Errorf("not implemented")
 }
 func (s priceStubExchange) EnsureOneWayMode() error { return nil }
 func (s priceStubExchange) Close()                  {}